refactor: provider, type 리팩터링
