+++ src/containers/random-pick/random-pick-container.tsx
@@ -4,8 +4,8 @@ import { useParams } from 'next/navigation';
 import useLocalStorage from '@/hooks/useLocalStorage';
 import PlayGround from './playground/playground';
 import RandomPickList from './random-pick-list/random-pick-list';
-import { RandomPickType } from './random-pick-provider/random-pick-provider';
 import RandomPickPlaygroundProvider from './random-pick-playground-provider.tsx/random-pick-playground-provider';
+import { RandomPickType } from './random-pick-type';
 
 export default function RandomPickContainer() {
   const { id } = useParams<{ id?: string }>();
